backend/python-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
requests==2.31.0
numpy==1.26.3
